Fix i3en.xlarge problem in scylla_io_setup.
The io numbers of i3en.xlarge that are calculated do not match with
what is measured:

Measured

    read_iops: 92776
    read_bandwidth: 748515456
    write_iops: 72220
    write_bandwidth: 330152736

Predicted:

    read_bandwidth: 2043674624
    read_iops: 257024
    write_bandwidth: 1024458752
    write_iops: 174080

The cause is there i3en.xlarge is not correctly written in the
i3en branch. It is written as i3.xlarge.

After fixing we get the following prediction:

    read_bandwidth: 666894336
    read_iops: 84480
    write_bandwidth: 333447168
    write_iops: 66969

Which is much more in line with what is beeing measured.