hints: introduce dynamic send concurrency limiter
Currently, hinted handoff uses a fixed (but configurable) per-node
concurrency limit - the memory they take up cannot exceed 10% of node's
memory and their count is bounded by the
`max_hinted_handoff_concurrency` option. Because hints reuse the
mutation sending framework and use the same timeout as regular writes,
if the hints concurrency is too much for the receiving node to handle,
hint writes will start to time out. In an event of a failed hint, the
sender waits for a second and retries sending after that delay. Due to
the relentlessness of this logic, sending hints might generate lots of
timeouts - which is wasted effort - or even get stuck in extreme cases.

In order to make current logic less aggressive and reduce the number
of timeouts, an adaptive strategy based on the concept of AIMD
(additive increase, multiplicative decrease) is introduced. Apart from
the current global limit, each hint queue gets an additional, dynamic
memory limit. The limit, similarly to the current one, allows to send
one hint at a time if a hint larger than the limit is encountered.

1. The limit starts with 1 byte.
2. Every second, the limit is increased by 10% of the size of an
   average hint. The average is calculated based on the hints sent
   since the last update.
3. In case of a timeout, the concurrency is cut in half, and the next
   update described in point 2 is scheduled to happen after 10 seconds.

The backoff behavior in case of a timeout is also changed:

1. If the concurrency == 1 byte, the next sending attempt is scheduled
   after a random delay between 1 and 120 seconds,
2. Otherwise the next sending attempt is scheduled after 1 second.