commitlog: Do a more exhaustive delete-drain on shutdown
Fixes #9303 (?)

Waiting for replensish _can_ iff we are super unlucky actually put
a new segment on reserve list. Need to delete/drain twice (or check
in replenish, but that is more intrusive)

Note: I could not repro this, as it will not happen very often.
But maybe the fix works...