cql3: expr: Implement evaluate(expr::usertype_constructor)
usertype_constructor can now be evaluated.

To evaluate an usertype_constructor we need to know the type,
because the fields have to be in the correct order.
Type has been added to usertype_constructor.

New code corresponds to old bind() of user_types::delayed_value in cql3/user_types.cc.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>