main: Sanitize resolving of broadcast address
To resolve this one main selects between the config option of
the same name or picks the listen address. Similarly to the
broadcast RPC address, on error the thrown exception is very
generic and doesn't tell which option contained the faulty
address.

THe utils::resolve, || operator and dedicated excplicit sanity
check make this place look better.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>