gossip: avoid dynamic call to snitch to get its name
Gossip has a safety check when joining a cluster where it compares
the node's snitch name to the rest of the cluster, and fails if they
don't match. This causes a circular dependency: gossip depends on
snitch to get its name, and snitch depends on gossip to pass along
dc/rack info.

To break the circle, initialize gossip with the snitch name during
construction.

This removes the ability for the snitch to change its name (via
reset_snitch()). However, that ability is useless - since gossip
explicity disallows clusters with mismatched snitches. If we ever
need that functionality, we can add an update_local_snitch_name()
method.