alternator: make three utility functions in executor.cc non-static
Make three of the utility functions in alternator/executor.cc, which
until now were static (local to the source files) external symbols
(in the alternator namespace). This will allow using them in other
Alternator source files - like the one in the next patch for TTL
support, which we'll want to put in a separate source file.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>