storage_service: coroutinize stop()
In preparation for adding more stuff, convert stop() to a coroutine
to avoid an unreadable chain of continuations.

The code uses a finally() block which might not be needed (since .join()
should not fail). Rather than risking getting it wrong I wrapped it in
a try/catch and added logging.